+++ CHANGELOG.md
@@ -11,7 +11,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ### Added
 
-- Allow constraining allowed relations in `@whereConditions` and `@whereHasConditions`
+- Allow constraining allowed relations in `@whereConditions` and `@whereHasConditions` https://github.com/nuwave/lighthouse/pull/1896
 
 ## v5.33.0